+++ tests/Unit/Evolution/Systems/CurvedScalarWave/Test_DuDt.cpp
@@ -13,7 +13,7 @@
 #include "Utilities/Gsl.hpp"
 #include "Utilities/MakeWithValue.hpp"
 #include "tests/Unit/Evolution/Systems/CurvedScalarWave/TestHelpers.hpp"
-#include "tests/Unit/PointwiseFunctions/GeneralRelativity/GrTestHelpers.hpp"
+#include "tests/Unit/PointwiseFunctions/GeneralRelativity/TestHelpers.hpp"
 
 // IWYU pragma: no_forward_declare Tensor
